+++ mdagent/agent/tools.py
@@ -3,10 +3,10 @@ import os
 from langchain import agents
 from langchain.base_language import BaseLanguageModel
 
-from ..general_tools import (
+from ..tools.gen_tools.search_tools import Scholar2ResultLLM
+from ..tools.md_utils.vis_tools import (
     CheckDirectoryFiles,
     PlanBVisualizationTool,
-    Scholar2ResultLLM,
     VisualizationToolRender,
 )
 
@@ -18,7 +18,13 @@ def make_tools(llm: BaseLanguageModel, verbose=False):
     all_tools = agents.load_tools(["python_repl", "human", "llm-math"], llm)
 
     # add visualization tools
-    all_tools += [CheckDirectoryFiles, VisualizationToolRender, PlanBVisualizationTool]
+    all_tools += [
+        CheckDirectoryFiles,
+        VisualizationToolRender,
+        PlanBVisualizationTool,
+    ]
+
+    # add literature search tool
     if pqa_key:
         all_tools.append(Scholar2ResultLLM(pqa_key))
     return all_tools
